add chatWrapper to app
